+++ infrastructure/src/main/java/mops/controllers/FragebogenRepository.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import mops.Fragebogen;
 import mops.SkalarFrage;
-import mops.TextFrage;
+import mops.fragen.TextFrage;
 
 
 public interface FragebogenRepository {
