+++ infrastructure/src/main/java/mops/database/MockFragebogenRepository.java
@@ -11,9 +11,9 @@ import mops.Auswahl;
 import mops.Einheit;
 import mops.Frage;
 import mops.Fragebogen;
-import mops.MultipleChoiceFrage;
+import mops.fragen.MultipleChoiceFrage;
 import mops.SkalarFrage;
-import mops.TextFrage;
+import mops.fragen.TextFrage;
 import mops.controllers.FragebogenRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Repository;
