+++ model/src/main/java/mops/antworten/Antwort.java
@@ -1,4 +1,4 @@
-package mops;
+package mops.antworten;
 
 @SuppressWarnings("PMD.AbstractClassWithoutAbstractMethod")
 public abstract class Antwort {
