+++ model/src/main/java/mops/antworten/MultipleResponseAntwort.java
@@ -1,8 +1,9 @@
-package mops;
+package mops.antworten;
 
 
 import java.util.HashSet;
 import java.util.Set;
+import mops.antworten.Antwort;
 
 public class MultipleResponseAntwort extends Antwort {
   transient Set<Integer> gewaehlteAntworten;
