+++ model/src/main/java/mops/antworten/TextAntwort.java
@@ -1,7 +1,8 @@
-package mops;
+package mops.antworten;
 
 import lombok.AllArgsConstructor;
 import lombok.Value;
+import mops.antworten.Antwort;
 
 @Value
 @AllArgsConstructor
