+++ model/src/main/java/mops/fragen/MultipleChoiceFrage.java
@@ -1,4 +1,4 @@
-package mops;
+package mops.fragen;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -6,6 +6,9 @@ import java.util.List;
 import java.util.Map;
 import lombok.Getter;
 import lombok.Setter;
+import mops.Auswahl;
+import mops.Frage;
+import mops.antworten.MultipleChoiceAntwort;
 
 @Getter
 @Setter
