+++ model/src/main/java/mops/fragen/MultipleResponseFrage.java
@@ -1,4 +1,6 @@
-package mops;
+package mops.fragen;
+
+import mops.fragen.MultipleChoiceFrage;
 
 public class MultipleResponseFrage extends MultipleChoiceFrage {
