+++ model/src/main/java/mops/fragen/SingleResponseFrage.java
@@ -1,4 +1,6 @@
-package mops;
+package mops.fragen;
+
+import mops.fragen.MultipleChoiceFrage;
 
 public class SingleResponseFrage extends MultipleChoiceFrage {
   final transient boolean isScala;
