+++ model/src/main/java/mops/fragen/TextFrage.java
@@ -1,9 +1,11 @@
-package mops;
+package mops.fragen;
 
 import java.util.HashSet;
 import java.util.Set;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
+import mops.Frage;
+import mops.antworten.TextAntwort;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
