+++ services/src/test/java/mops/SubmitServiceTest.java
@@ -9,6 +9,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import mops.antworten.TextAntwort;
+import mops.fragen.TextFrage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
