+++ services/src/test/java/mops/TypeCheckerTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import mops.fragen.MultipleChoiceFrage;
+import mops.fragen.TextFrage;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings( {"PMD.JUnitTestsShouldIncludeAssert", "PMD.JUnitAssertionsShouldIncludeMessage"})
